+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 0.20
 
 * #89 Don't print debug info to stdout
+* #91 bootique-test: test helper for PolymorphicConfiguration
 
 ## 0.19